backend/pyproject.toml
[project]
name = "gitthub-backend"
version = "0.1.0"
description = "FastAPI backend for gitthub.org - serves React frontend and manages workflows"
requires-python = ">=3.12"
dependencies = [
    # Local packages
    "anthropic-api @ file:///Users/gitt/hub/anthropic-api",

    # FastAPI and Web Server
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",

    # Database
    "sqlalchemy>=2.0.23",
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",
    "alembic>=1.12.1",

    # Settings and Environment
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "email-validator>=2.1.0",

    # Authentication and Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.1.2",
    "python-multipart>=0.0.6",
    "PyJWT>=2.8.0",

    # AWS SDK
    "boto3>=1.34.0",

    # Utilities
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["app"]